% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrace_functions.R
\docType{package}
\name{cloudtrace_googleAuthR}
\alias{cloudtrace_googleAuthR}
\alias{cloudtrace_googleAuthR-package}
\title{Stackdriver Trace API
Send and retrieve trace data from Stackdriver Trace. Data is generated and available by default for all App Engine applications. Data from other applications can be written to Stackdriver Trace for display, reporting, and analysis.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:34:03
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudtracev1.auto/R/cloudtrace_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/trace.readonly
\item https://www.googleapis.com/auth/trace.append
}
}
