% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrace_objects.R
\name{ListTracesResponse}
\alias{ListTracesResponse}
\title{Google Cloud Trace API Objects 
Send and retrieve trace data from Google Cloud Trace. Data is generated and available by default for all App Engine applications. Data from other applications can be written to Cloud Trace for display, reporting, and analysis.}
\usage{
ListTracesResponse(traces = NULL, nextPageToken = NULL)
}
\arguments{
\item{traces}{List of trace records returned}

\item{nextPageToken}{If defined, indicates that there are more traces that match the request and that this value should be passed to the next request to continue retrieving additional traces}
}
\value{
ListTracesResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:17:45
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudtracev1.auto/R/cloudtrace_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListTracesResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for the `ListTraces` method.
}

