% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrace_objects.R
\name{Traces}
\alias{Traces}
\title{Traces Object}
\usage{
Traces(traces = NULL)
}
\arguments{
\item{traces}{List of traces}
}
\value{
Traces object
}
\description{
Traces Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List of new or updated traces.
}
\seealso{
Other Traces functions: \code{\link{projects.patchTraces}}
}
