% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrace_objects.R
\name{TraceSpan.labels}
\alias{TraceSpan.labels}
\title{TraceSpan.labels Object}
\usage{
TraceSpan.labels()
}
\value{
TraceSpan.labels object
}
\description{
TraceSpan.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Collection of labels associated with the span. Label keys must be less than128 bytes. Label values must be less than 16 kilobytes.
}
\seealso{
Other TraceSpan functions: \code{\link{TraceSpan}}
}
