% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrace_objects.R
\name{Trace}
\alias{Trace}
\title{Trace Object}
\usage{
Trace(projectId = NULL, traceId = NULL, spans = NULL)
}
\arguments{
\item{projectId}{Project ID of the Cloud project where the trace data is stored}

\item{traceId}{Globally unique identifier for the trace}

\item{spans}{Collection of spans in the trace}
}
\value{
Trace object
}
\description{
Trace Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A trace describes how long it takes for an application to perform an operation. It consists of a set of spans, each of which represent a single timed event within the operation.
}

