% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrace_objects.R
\name{TraceSpan}
\alias{TraceSpan}
\title{TraceSpan Object}
\usage{
TraceSpan(TraceSpan.labels = NULL, spanId = NULL, name = NULL,
  startTime = NULL, endTime = NULL, parentSpanId = NULL, labels = NULL)
}
\arguments{
\item{TraceSpan.labels}{The \link{TraceSpan.labels} object or list of objects}

\item{spanId}{Identifier for the span}

\item{name}{Name of the trace}

\item{startTime}{Start time of the span in nanoseconds from the UNIX epoch}

\item{endTime}{End time of the span in nanoseconds from the UNIX epoch}

\item{parentSpanId}{ID of the parent span, if any}

\item{labels}{Collection of labels associated with the span}
}
\value{
TraceSpan object
}
\description{
TraceSpan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A span represents a single timed event within a trace. Spans can be nested and form a trace tree. Often, a trace contains a root span that describes the end-to-end latency of an operation and, optionally, one or more subspans for its suboperations. Spans do not need to be contiguous. There may be gaps between spans in a trace.
}
\seealso{
Other TraceSpan functions: \code{\link{TraceSpan.labels}}
}

